requirements.txt
playwright~=1.41.2
pytest~=8.0.2
allure-pytest